compile on linux

+++ src/fluxions_gte_spherical_harmonic.cpp
@@ -1251,7 +1251,7 @@ namespace Fluxions
 	bool Sph4f::fromVectorFormat(int maxDegrees, const std::vector<std::vector<float>> & v)
 	{
 		// do a sanity check on the incoming data
-		int lmCount = maxDegree * (maxDegree + 1) + maxDegree + 1;
+		size_t lmCount = maxDegree * (maxDegree + 1) + maxDegree + 1;
 		if (lmCount != v.size()) return false;
 
 		// then attempt to resize to fit the incoming data
